Adiciona testes automatizados e configurações para ambiente de testes
- Cria testes para a classe FormatadorData, validando o parsing de datas.
- Implementa testes para SensitiveDataSanitizer, assegurando a mascaramento de dados sensíveis.
- Adiciona testes para ExecutionContext, verificando a inicialização e propagação do contexto.
- Inclui arquivos de exemplo para testes de integração com GraphQL e dados de manifesto.
- Configura ambiente de testes com Docker, incluindo serviços de SQL Server e Wiremock.
- Adiciona datasets sintéticos para simular cenários de erro, duplicatas, e paginação incompleta.

+++ database/tabelas/012_criar_tabela_sys_execution_history.sql
@@ -5,7 +5,7 @@
 IF NOT EXISTS (SELECT * FROM sys.objects WHERE object_id = OBJECT_ID(N'dbo.sys_execution_history') AND type in (N'U'))
 BEGIN
     CREATE TABLE dbo.sys_execution_history (
-        id INT IDENTITY(1,1) PRIMARY KEY,
+        id BIGINT IDENTITY(1,1) PRIMARY KEY,
         start_time DATETIME NOT NULL,
         end_time DATETIME NOT NULL,
         duration_seconds INT NOT NULL,
